App/Relay/DaylightDetector/requirements.txt
smbus
RPi.GPIO
